requirements.txt
# ============================================================
# requirements.txt — football_defence_analysis
# Install with: pip install -r requirements.txt
# ============================================================

# --- API & HTTP ---
requests==2.31.0
python-dotenv==1.0.0

# --- Data Handling ---
pandas==2.2.0
numpy==1.26.4

# --- Statistical Analysis ---
scipy==1.12.0
statsmodels==0.14.1

# --- Machine Learning ---
scikit-learn==1.4.0

# --- Visualization ---
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0

# --- Notebooks ---
jupyterlab==4.1.0
ipykernel==6.29.0

# --- Utilities ---
tqdm==4.66.1
colorlog==6.8.2